mysite/requirements.txt
django==4.1.2
django-allauth==0.51.0
requests==2.28.1
requests-oauthlib==1.3.1
oauthlib==3.2.1
requests-oauthlib==1.3.1